refactor: move js code away from copytoclip python file
